overhaul test result reporting, exit codes
tests can now result in success, skip, failure, or breakage/error,
with a distinct report for each.  broken tests are labeled "X".

a test is broken when its exit code is a non-zero different
from 80 (80 is skip).  cram does not handle such tests well::

  $ cat > testfile <<\CRAM
  >   $ exit 1
  >   $ echo foo
  > CRAM
  $ cram testfile
  .
  # Ran 1 tests, 0 skipped, 0 failed.

btw, dram does not print diffs for broken tests.

summary now has a slightly different format, and dynamic contents:
the number of skipped tests is reported only if it's non-zero;
mutatis mutandis for failures and errors.

if there were failures, dram exits 2.  if there were errors,
the exit code is 4.  if there were both, it's 6.